+++ src/web/src/components/organisms/DocumentTable.tsx
@@ -2,7 +2,7 @@ import { useState, useMemo } from "react";
 import { semantic as t, useInjectStyles } from "@4lt7ab/ui/core";
 import { Icon } from "../atoms/Icon";
 import { IconButton } from "../atoms/IconButton";
-import { TagChip } from "../molecules/TagChip";
+import { TagChip } from "@4lt7ab/ui/ui";
 import { useTheme } from "../theme/ThemeContext";
 import { useWindowWidth, SMALL_BREAKPOINT } from "../../hooks/useWindowWidth";
 import type { DocumentSummary } from "../../types";